requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
pytest
pytest-asyncio